Reuse tellor 360 test contracts setup
