+++ tests/conftest.py
@@ -3,6 +3,7 @@ import asyncio
 import os
 
 import pytest
+import pytest_asyncio
 from brownie import accounts
 from brownie import Autopay
 from brownie import chain
@@ -18,6 +19,7 @@ from chained_accounts import find_accounts
 from multicall import multicall
 from multicall.constants import MULTICALL2_ADDRESSES
 from multicall.constants import Network
+from telliot_core.apps.core import TelliotCore
 from telliot_core.apps.telliot_config import TelliotConfig
 
 from telliot_feeds.datasource import DataSource
@@ -281,3 +283,32 @@ def tellorflex_360_contract(mock_token_contract):
         1,
         "0x5c13cd9c97dbb98f2429c101a2a8150e6c7a0ddaff6124ee176a3a411067ded0",
     )
+
+
+@pytest_asyncio.fixture(scope="function")
+async def tellor_360(mumbai_test_cfg, tellorflex_360_contract, mock_autopay_contract, mock_token_contract):
+    async with TelliotCore(config=mumbai_test_cfg) as core:
+        txn_kwargs = {"gas_limit": 3500000, "legacy_gas_price": 1}
+        account = core.get_account()
+
+        tellor360 = core.get_tellor360_contracts()
+        tellor360.oracle.address = tellorflex_360_contract.address
+        tellor360.oracle.abi = tellorflex_360_contract.abi
+        tellor360.autopay.address = mock_autopay_contract.address
+        tellor360.autopay.abi = mock_autopay_contract.abi
+        tellor360.token.address = mock_token_contract.address
+
+        tellor360.oracle.connect()
+        tellor360.token.connect()
+        tellor360.autopay.connect()
+
+        # mint token and send to reporter address
+        mock_token_contract.mint(account.address, 100000e18)
+
+        # send eth from brownie address to reporter address for txn fees
+        accounts[1].transfer(account.address, "1 ether")
+
+        # init governance address
+        await tellor360.oracle.write("init", _governanceAddress=accounts[0].address, **txn_kwargs)
+
+        return tellor360, account
